[fix] 골드를 long 타입으로 바꾼 것을 다른 코드에 적용

+++ Assets/02_Scripts/Data/SaveData.cs
@@ -4,5 +4,5 @@
 public class SaveData
 {
     public int CurrentStage = 1;
-    public float Gold = 0;
+    public long Gold = 0;
 }
